roles: remove Balloonist "Seen Traveller" reminder
Though I agree showing a Traveler to a Balloonist before N5 is bad
storytelling, I think the token text obligates it, unless we are to
believe that Travelers aren't a character type.

However, Edd has ruled they are not, in fact, a character type:
  https://discord.com/channels/569683781800296501/570328029939630090/696808482799616021

I personally feel this is ad-hoc to fix Balloonist when really the
Balloonist token text needs an "(except Travelers)".

This isn't an opinionated fork (yet), so we go with the official
ruling and remove the reminder token.